Fix some errors in responses that makes errors in client side.

+++ Fleet ManagementWebApplication/Models/RouteHistory.cs
@@ -7,7 +7,7 @@
         public int VehicleDirection { get; set; }
         public char Status { get; set; }
         public string VehicleSpeed { get; set; }
-        public int Epoch { get; set; }
+        public long Epoch { get; set; }
         public string Address { get; set; }
         public float Latitude { get; set; }
         public float Longitude { get; set; }
